backend/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pytesseract==0.3.13
Werkzeug==3.0.4
